Spring programmatic transaction management

+++ src/com/izdebski/model/Account.java
@@ -2,39 +2,32 @@ package com.izdebski.model;
 
 public class Account {
 
-    private Long accountNumber;
+    private Integer accountNumber;
     private Double accountBalance;
     private String accounType;
     private String accountHolderName;
 
-    public Long getAccountNumber() {
+    public Integer getAccountNumber() {
         return accountNumber;
     }
-
-    public void setAccountNumber(Long accountNumber) {
+    public void setAccountNumber(Integer accountNumber) {
         this.accountNumber = accountNumber;
     }
-
     public Double getAccountBalance() {
         return accountBalance;
     }
-
     public void setAccountBalance(Double accountBalance) {
         this.accountBalance = accountBalance;
     }
-
     public String getAccounType() {
         return accounType;
     }
-
     public void setAccounType(String accounType) {
         this.accounType = accounType;
     }
-
     public String getAccountHolderName() {
         return accountHolderName;
     }
-
     public void setAccountHolderName(String accountHolderName) {
         this.accountHolderName = accountHolderName;
     }
